+++ man-roxygen/param_store_models.R
@@ -1,4 +1,5 @@
 #' @param store_models (`logical(1)`)\cr
-#'   Store the fitted model in the resulting object=
+#'   Store the fitted model in the resulting object.
 #'   Set to `TRUE` if you want to further analyze the models or want to
 #'   extract information like variable importance.
+#'   If `FALSE`, the model of the fallback learner is discarded as well.